requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.2.0
gunicorn>=21.2.0
cryptography>=41.0.0
Flask-Babel>=4.0.0
